requirements.txt
Flask
qrcode
opencv-python
pyzbar
Pillow
numpy
